requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
numpy==1.26.0
Pillow==10.0.1
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.1
matplotlib==3.8.0
pandas==2.1.1
opencv-python==4.8.1.78
albumentations==1.3.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
pydantic==2.4.2
aiofiles==23.2.1
pytest==7.4.2
email-validator 